Update Program.cs
Vol.2 (Comments was not saved :D)

Program is pretty easy to read but i added some comments to ease it.

It could always be a bit more modular or bit more something but now it works and in real life i think its important get job done.

If you are reading this and  you have some better solution or suggest to some functions please leave a comment :) I would appreciate it!